feat: reduce package size with platform specific distributions (#38)
Reduces package size by a factor of 6: ~66MB to ~11MB.

Changes how the CLI binaries are added to the python package, and how
the python executable launches the binary.

1. The binaries are now located in `numerous/cli/bin`.
2. The python script tries to launch `numerous/cli/bin/cli` if it
   exists. It exists if the package is installed from a platform
   specific wheel.
3. Otherwise the python script will attempt to match the the current OS
   with a binary, similar to the approach until now. This is meant for
   non-platform specific wheels.

The build process is changed. Wheels and source distribution are created
by `scripts/build_dists.sh`, which can create a bdist for each relevant
platform that we build CLI binaries for. The CLI binary directory
`numerous/cli/bin` is cleared, and the relevant CLI binary is copied
into `numerous/cli/bin/cli`, and the corresponding wheel is created.

For the platform non-specific wheel and the source distribution all
binaries are added, as before.